requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
pydantic>=2.7
pydantic-settings>=2.2
python-dotenv>=1.0
prometheus-fastapi-instrumentator>=6.1.0
itsdangerous>=2.2
passlib[bcrypt]>=1.7
redis>=5
Jinja2>=3.1
PyJWT>=2.8
